+++ taskmonitor/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "taskmonitor.apps.TaskMonitorConfig"
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"
 __title__ = "Task Monitor"